requirements.txt
sentence-transformers==2.6.0
faiss-cpu==1.7.4
langgraph==0.0.67
langchain==0.1.20
langchain-ollama==0.1.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.0
numpy==1.24.3
